Implementation of threads into sensor (unsure if well done) and updated cmake so variables can be declared in loops.

+++ light_sampler/hal/include/hal/timing.h
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <time.h>
 
-int nanotoms (int ns);
+double nanotoms (int ns);
 
 
 long long getTimeInMs(void);
